+++ libs/api/a-state/src/lib/products/dto/index.ts
@@ -1,3 +1,4 @@
 export * from './create-products.dto';
 export * from './generate-product-pdf.dto';
+export * from './get-products.dto';
 export * from './update-product.dto';