+++ libs/ui/u-groups/src/lib/components/assign-products-modal/assign-products-modal.component.ts
@@ -1,0 +1,35 @@
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnChanges, Output } from '@angular/core';
+import { Product } from '@ivt/c-data';
+import { TransferItem } from 'ng-zorro-antd/transfer';
+
+@Component({
+  selector: 'ivt-assign-products-modal',
+  templateUrl: './assign-products-modal.component.html',
+  styleUrls: ['./assign-products-modal.component.less'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class AssignProductsModalComponent implements OnChanges {
+  @Input() products: Product[];
+  @Input() groupProducts: Product[];
+  transferData: TransferItem[] = [];
+  @Output() submitData = new EventEmitter<number[]>();
+
+  get loading(): boolean {
+    return !this.products || !this.groupProducts;
+  }
+
+  ngOnChanges() {
+    if (this.products && this.groupProducts) {
+      this.transferData = this.products.map(product => ({
+        key: product.id,
+        title: product.name,
+        description: product.logo,
+        direction: this.groupProducts.find(groupProduct => groupProduct.id === product.id) ? 'right' : undefined,
+      }));
+    }
+  }
+
+  submit(): void {
+    this.submitData.emit(this.transferData.filter(item => item.direction === 'right').map(item => item.key));
+  }
+}