+++ libs/ui/u-groups/src/lib/containers/group-list-container/group-list-container.component.ts
@@ -2,6 +2,9 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { Group } from '@ivt/c-data';
 import { GroupCollectionService, GroupDataService } from '@ivt/u-state';
 import { IvtListContainerComponent } from '@ivt/u-ui';
+import { NzModalService } from 'ng-zorro-antd/modal';
+
+import { AssignProductsModalContainerComponent } from '../assign-products-modal-container/assign-products-modal-container.component';
 
 @Component({
   selector: 'ivt-group-list-container',
@@ -10,7 +13,21 @@ import { IvtListContainerComponent } from '@ivt/u-ui';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class GroupListContainerComponent extends IvtListContainerComponent<Group> {
-  constructor(protected groupCollectionService: GroupCollectionService, protected groupDataService: GroupDataService) {
+  constructor(
+    protected groupCollectionService: GroupCollectionService,
+    protected groupDataService: GroupDataService,
+    private modalService: NzModalService
+  ) {
     super(groupCollectionService, groupDataService);
+  }
+
+  assignProducts(group: Group): void {
+    this.modalService.create({
+      nzTitle: `Asignar productos - ${group.name}`,
+      nzContent: AssignProductsModalContainerComponent,
+      nzComponentParams: { groupId: group.id },
+      nzStyle: { minWidth: '85vw' },
+      nzOnOk: component => component.submitChild(),
+    });
   }
 }