+++ libs/ui/u-vehicles/src/lib/public/vehicle-form/vehicle-form.component.ts
@@ -14,6 +14,7 @@ import { IvtFormComponent } from '@ivt/u-ui';
 import { QueryParams } from '@ngrx/data';
 import { omit } from 'lodash-es';
 import { NzSelectOptionInterface } from 'ng-zorro-antd/select';
+import { combineLatest } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
 
 export function createVehicleForm(vehicle?: Vehicle) {
@@ -31,7 +32,7 @@ export function createVehicleForm(vehicle?: Vehicle) {
       ApsValidators.maxLength(VEHICLE_VIN_LENGTH),
     ]),
     motorNumber: new FormControl(null),
-    horsePower: new FormControl(null, [ApsValidators.requiredNumber, ApsValidators.min(1), ApsValidators.max(400)]),
+    horsePower: new FormControl(null, [ApsValidators.requiredNumber, ApsValidators.min(1), ApsValidators.max(500)]),
     companyId: new FormControl(null, ApsValidators.required),
   });
 
@@ -56,6 +57,7 @@ export class VehicleFormComponent extends IvtFormComponent<Vehicle> implements O
   @Input() companiesInfo: IvtCollectionResponseInfo;
   @Output() verifyVin = new EventEmitter<string>();
   @Output() getCompanies = new EventEmitter<QueryParams>();
+  @Output() getProducts = new EventEmitter<{ year: string; horsePower: string }>();
 
   ngOnChanges(changes: SimpleChanges) {
     if (changes.companyId) {
@@ -70,6 +72,10 @@ export class VehicleFormComponent extends IvtFormComponent<Vehicle> implements O
           takeUntil(this.destroy$)
         )
         .subscribe(vin => this.verifyVin.emit(vin));
+
+      combineLatest([this.form.get('horsePower').valueChanges, this.form.get('year').valueChanges])
+        .pipe(takeUntil(this.destroy$))
+        .subscribe(([horsePower, year]) => this.getProducts.emit({ horsePower, year }));
     }
 
     if (changes.showCompanyInput) {
@@ -82,6 +88,7 @@ export class VehicleFormComponent extends IvtFormComponent<Vehicle> implements O
       this.form.patchValue(this.item, { emitEvent: false });
       this.form.disable({ emitEvent: false });
       this.form.get('vin').enable({ emitEvent: false });
+      this.getProducts.emit({ horsePower: String(this.item.horsePower), year: String(this.item.year) });
     }
 
     if (changes.vehicle && this.isEditable) {
@@ -89,6 +96,7 @@ export class VehicleFormComponent extends IvtFormComponent<Vehicle> implements O
         this.form.patchValue(omit(this.vehicle, 'vin'));
         this.form.disable({ emitEvent: false });
         this.form.get('vin').enable({ emitEvent: false });
+        this.getProducts.emit({ horsePower: String(this.vehicle.horsePower), year: String(this.vehicle.year) });
       } else {
         this.form.enable({ emitEvent: false });
         this.form.get('id').patchValue(null);